feat: extract shared templates crate and unify error mapping
- Create statespace-templates crate with AGENTS_MD, FAVICON_SVG, and
  render_index_html, shared between OSS and hosted servers
- Add Error::http_status_code() to statespace-tool-runtime as single
  source of truth for error-to-HTTP-status mapping
- Update statespace-server to re-export from shared crate and delegate
  error mapping to the new method

Amp-Thread-ID: https://ampcode.com/threads/T-019c8831-5ec3-7253-abaa-e35ab0d4b7b8
